change the type of "ave_retrans_times_" from int to double

+++ satellite/RetransTimes.h
@@ -12,7 +12,7 @@ class RetransTimes : public TclObject
 public:
 	static int success_retrans_times_sum_;
 	static int success_pkt_num_;
-	static int ave_retrans_times_;
+	static double ave_retrans_times_;
 	static int k_;
 	static int prac_sent_pkt_;
